Fix remaining Kotlin compile errors
- Fix type mismatch in MainActivity.calculateSpanCount by using Float division.
- Replace ViewParent.width with casting to View and using measuredWidth.
- Use PlayerView.isControllerFullyVisible instead of non-existent isControllerVisible.

+++ app/src/main/java/com/example/tvplayer/ui/MainActivity.kt
@@ -73,7 +73,7 @@ class MainActivity : AppCompatActivity() {
     private fun calculateSpanCount(): Int {
         val displayMetrics = resources.displayMetrics
         val dpWidth = displayMetrics.widthPixels / displayMetrics.density
-        return (dpWidth / 320).coerceAtLeast(2)
+        return (dpWidth / 320f).coerceAtLeast(2f).toInt()
     }
 
     private fun setupInput() {
